Add alreadyKnownTagger, move browser functionality to browser folder
