+++ morph/browser/__init__.py
@@ -1,0 +1,1 @@
+version = 3.6
