+++ morph/browser/viewMorphemes.py
@@ -1,7 +1,7 @@
 #-*- coding: utf-8 -*-
-from .morphemes import getMorphemes, ms2str
-from .morphemizer import getMorphemizerByName
-from .util import addBrowserNoteSelectionCmd, getFilter, infoMsg
+from ..morphemes import getMorphemes, ms2str
+from ..morphemizer import getMorphemizerByName
+from ..util import addBrowserNoteSelectionCmd, getFilter, infoMsg
 
 def pre( b ): return { 'morphemes': [] }
